+++ src/CommandSeparator.vue
@@ -1,4 +1,8 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+defineProps<{
+  alwaysRender?: boolean
+}>()
+</script>
 
 <template>
   <div data-cmdk-separator="" role="separator" />